pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Homely Assist", email = "homeconnecteasy@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.50", # AWS Stuff (S3, Dynamo ...)
    "flask>=3.1.2", # REST Eedpoint, fast-api
    "google-generativeai>=0.8.5", # AI
    "langchain>=0.3.27", # framework for building ai chat
    "langgraph>=0.6.10", # framework for building ai agent
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1", # utils this is tool that load .env file
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
